+++ app/src/main/java/com/example/myapplication/LoginActivity.java
@@ -1,0 +1,120 @@
+package com.example.myapplication;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Patterns;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.ProgressBar;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+
+public class LoginActivity extends AppCompatActivity implements View.OnClickListener {
+
+    private TextView register, forgotPassword;
+    private EditText editTextEmail, editTextPassword;
+    private Button signIn;
+
+    private FirebaseAuth mAuth;
+    private ProgressBar progressBar;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        mAuth = FirebaseAuth.getInstance();
+
+        register = (TextView) findViewById(R.id.register);
+        register.setOnClickListener(this);
+
+        forgotPassword = (TextView) findViewById(R.id.forgotPassword);
+        forgotPassword.setOnClickListener(this);
+
+        signIn = (Button)findViewById(R.id.signIn);
+        signIn.setOnClickListener(this);
+
+        editTextEmail = (EditText)findViewById(R.id.email);
+        editTextPassword=(EditText)findViewById(R.id.password);
+
+        progressBar = (ProgressBar) findViewById(R.id.progressBar);
+    }
+
+    @Override
+    public void onClick(View v) {
+        switch (v.getId()) {
+            // What to do when the user presses any of the three buttons
+            case R.id.register:
+                startActivity(new Intent(this, RegisterUser.class));
+                break;
+            case R.id.signIn:
+                userLogin();
+                break;
+            case R.id.forgotPassword:;
+                startActivity(new Intent(this, com.example.myapplication.ForgotPassword.class));
+                break;
+        }
+    }
+
+    // Login using fireBase
+    private void userLogin() {
+        String email = editTextEmail.getText().toString().trim();
+        String password = editTextPassword.getText().toString().trim();
+
+        boolean validLogin = validate(email, password);
+        if(validLogin) {
+            progressBar.setVisibility(View.VISIBLE);
+            mAuth.signInWithEmailAndPassword(email, password).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                @Override
+                public void onComplete(@NonNull Task<AuthResult> task) {
+                    if(task.isSuccessful()) {
+                        //redirect to user profile
+                        startActivity(new Intent(LoginActivity.this, MainActivity.class));
+                    }
+                    else {
+                        Toast.makeText(LoginActivity.this, "Failed to login, please check your credentials.", Toast.LENGTH_LONG).show();
+                    }
+                    progressBar.setVisibility(View.GONE);
+                }
+            });
+        }
+    }
+
+    // Validate email, password, and password length
+    private boolean validate(String email, String password) {
+
+        if(email.isEmpty()) {
+            editTextEmail.setError("Email is required");
+            editTextEmail.requestFocus();
+            return false;
+        }
+        if(!Patterns.EMAIL_ADDRESS.matcher(email).matches()) {
+            editTextEmail.setError("Please enter a valid email");
+            editTextEmail.requestFocus();
+            return false;
+        }
+        if(password.isEmpty()) {
+            editTextPassword.setError("Password is required");
+            editTextPassword.requestFocus();
+            return false;
+        }
+        if(password.length() < 6) {
+            editTextPassword.setError("Password should contain a minimum of 6 characters.");
+            editTextPassword.requestFocus();
+            return false;
+        }
+
+        return true;
+
+    }
+
+}
