+++ app/src/main/java/com/example/myapplication/RegisterUser.java
@@ -50,7 +50,7 @@ public class RegisterUser extends AppCompatActivity implements View.OnClickListe
     public void onClick(View v) {
         switch(v.getId()) {
             case R.id.banner:
-                startActivity(new Intent(this, com.example.myapplication.MainActivity.class));
+                startActivity(new Intent(this, LoginActivity.class));
                 break;
             case R.id.registerUser:
                 registerUser();
@@ -58,6 +58,10 @@ public class RegisterUser extends AppCompatActivity implements View.OnClickListe
 
         }
 
+    }
+
+    public void onClickBack(View view){
+        this.finish();
     }
 
     private void registerUser() {
