+++ app/src/main/res/layout/activity_register_user.xml
@@ -7,6 +7,17 @@
     tools:context=".RegisterUser">
 
 
+    <Button
+        android:id="@+id/back"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginLeft="10dp"
+        android:layout_marginTop="10dp"
+        android:onClick="onClickBack"
+        android:backgroundTint="#0ED6B9"
+        android:padding="5dp"
+        android:text="Back" />
+
     <TextView
         android:id="@+id/banner"
         android:layout_width="match_parent"
